+++ interfaces/native/include/sensor_algorithm.h
@@ -23,27 +23,27 @@ class SensorAlgorithm {
 public:
     SensorAlgorithm() = default;
     ~SensorAlgorithm() = default;
-    int32_t CreateQuaternion(std::vector<float> rotationVector, std::vector<float> &quaternion);
-    int32_t TransformCoordinateSystem(std::vector<float> inRotationMatrix, int32_t axisX,
-                                    int32_t axisY, std::vector<float> &outRotationMatrix);
-    int32_t GetAltitude(float seaPressure, float currentPressure, float *altitude);
-    int32_t GetGeomagneticDip(std::vector<float> inclinationMatrix, float *geomagneticDip);
-    int32_t GetAngleModify(std::vector<float> currotationMatrix, std::vector<float> prerotationMatrix,
-                        std::vector<float> &angleChange);
-    int32_t GetDirection(std::vector<float> rotationMatrix, std::vector<float> &rotationAngle);
-    int32_t CreateRotationMatrix(std::vector<float> rotationVector, std::vector<float> &rotationMatrix);
-    int32_t CreateRotationAndInclination(std::vector<float> gravity, std::vector<float> geomagnetic,
-                                        std::vector<float> &rotationMatrix, std::vector<float> &inclinationMatrix);
+    int32_t CreateQuaternion(std::vector<double> rotationVector, std::vector<double> &quaternion);
+    int32_t TransformCoordinateSystem(std::vector<double> inRotationMatrix, int32_t axisX,
+                                    int32_t axisY, std::vector<double> &outRotationMatrix);
+    int32_t GetAltitude(double seaPressure, double currentPressure, double *altitude);
+    int32_t GetGeomagneticDip(std::vector<double> inclinationMatrix, double *geomagneticDip);
+    int32_t GetAngleModify(std::vector<double> currotationMatrix, std::vector<double> prerotationMatrix,
+                        std::vector<double> &angleChange);
+    int32_t GetDirection(std::vector<double> rotationMatrix, std::vector<double> &rotationAngle);
+    int32_t CreateRotationMatrix(std::vector<double> rotationVector, std::vector<double> &rotationMatrix);
+    int32_t CreateRotationAndInclination(std::vector<double> gravity, std::vector<double> geomagnetic,
+                                        std::vector<double> &rotationMatrix, std::vector<double> &inclinationMatrix);
 
 private:
-    int32_t TransformCoordinateSystemImpl(std::vector<float> inRotationMatrix, int32_t axisX,
-                                                           int32_t axisY, std::vector<float> &outRotationMatrix);
+    int32_t TransformCoordinateSystemImpl(std::vector<double> inRotationMatrix, int32_t axisX,
+                                                           int32_t axisY, std::vector<double> &outRotationMatrix);
     static constexpr int32_t QUATERNION_LENGTH = 4;
     static constexpr int32_t ROTATION_VECTOR_LENGTH = 3;
     static constexpr int32_t THREE_DIMENSIONAL_MATRIX_LENGTH = 9;
     static constexpr int32_t FOUR_DIMENSIONAL_MATRIX_LENGTH = 16;
-    static constexpr float GRAVITATIONAL_ACCELERATION = 9.81f;
-    static constexpr float RECIPROCAL_COEFFICIENT = 5.255f;
-    static constexpr float ZERO_PRESSURE_ALTITUDE = 44330.0f;
+    static constexpr double GRAVITATIONAL_ACCELERATION = 9.81f;
+    static constexpr double RECIPROCAL_COEFFICIENT = 5.255f;
+    static constexpr double ZERO_PRESSURE_ALTITUDE = 44330.0f;
 };
 #endif // SENSOR_ALGORITHM_H
